# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SEO Local Admin Panel API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface AdminUser {
  id: number;
  email: string;
  createdAt?: string;
}

export interface AuthResponse {
  token: string;
  user: AdminUser;
}

export interface DashboardStats {
  totalClients: number;
  activeClients?: number;
  activeCampaigns: number;
  totalKeywords: number;
  newBacklinks30d: number;
  totalRevenue?: number;
}

export interface ActivityItem {
  id: number;
  type: string;
  description: string;
  /** @nullable */
  entityId?: number | null;
  createdAt: string;
}

export interface RevenueDataPoint {
  month: string;
  revenue: number;
}

export interface KeywordAlert {
  keywordId: number;
  keywordText: string;
  campaignName: string;
  clientName: string;
  /** @nullable */
  previousRank: number | null;
  /** @nullable */
  currentRank: number | null;
  drop: number;
}

export type ClientStatus = typeof ClientStatus[keyof typeof ClientStatus];


export const ClientStatus = {
  active: 'active',
  inactive: 'inactive',
} as const;

export interface Client {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  company?: string | null;
  /** @nullable */
  websiteUrl?: string | null;
  /** @nullable */
  assignedPlanId?: number | null;
  /** @nullable */
  planName?: string | null;
  status: ClientStatus;
  createdAt: string;
  updatedAt?: string;
}

export type ClientInputStatus = typeof ClientInputStatus[keyof typeof ClientInputStatus];


export const ClientInputStatus = {
  active: 'active',
  inactive: 'inactive',
} as const;

export interface ClientInput {
  name: string;
  email: string;
  company?: string;
  websiteUrl?: string;
  assignedPlanId?: number;
  status?: ClientInputStatus;
}

export type ClientUpdateStatus = typeof ClientUpdateStatus[keyof typeof ClientUpdateStatus];


export const ClientUpdateStatus = {
  active: 'active',
  inactive: 'inactive',
} as const;

export interface ClientUpdate {
  name?: string;
  email?: string;
  company?: string;
  websiteUrl?: string;
  /** @nullable */
  assignedPlanId?: number | null;
  status?: ClientUpdateStatus;
}

export interface ClientsPage {
  clients: Client[];
  total: number;
  page: number;
  limit: number;
}

export interface Plan {
  id: number;
  name: string;
  price: number;
  keywordLimit: number;
  backlinkCheckLimit: number;
  /** @nullable */
  auditLimit?: number | null;
  createdAt: string;
}

export interface PlanInput {
  name: string;
  price: number;
  keywordLimit: number;
  backlinkCheckLimit: number;
  auditLimit?: number;
}

export interface PlanUpdate {
  name?: string;
  price?: number;
  keywordLimit?: number;
  backlinkCheckLimit?: number;
  /** @nullable */
  auditLimit?: number | null;
}

export type CampaignStatus = typeof CampaignStatus[keyof typeof CampaignStatus];


export const CampaignStatus = {
  active: 'active',
  paused: 'paused',
  completed: 'completed',
} as const;

export interface Campaign {
  id: number;
  clientId: number;
  /** @nullable */
  clientName?: string | null;
  name: string;
  targetDomain: string;
  /** @nullable */
  targetLocation?: string | null;
  /** @nullable */
  targetLanguage?: string | null;
  status: CampaignStatus;
  keywordCount?: number;
  createdAt: string;
}

export type CampaignInputStatus = typeof CampaignInputStatus[keyof typeof CampaignInputStatus];


export const CampaignInputStatus = {
  active: 'active',
  paused: 'paused',
  completed: 'completed',
} as const;

export interface CampaignInput {
  clientId: number;
  name: string;
  targetDomain: string;
  targetLocation?: string;
  targetLanguage?: string;
  status?: CampaignInputStatus;
}

export type CampaignUpdateStatus = typeof CampaignUpdateStatus[keyof typeof CampaignUpdateStatus];


export const CampaignUpdateStatus = {
  active: 'active',
  paused: 'paused',
  completed: 'completed',
} as const;

export interface CampaignUpdate {
  name?: string;
  targetDomain?: string;
  /** @nullable */
  targetLocation?: string | null;
  /** @nullable */
  targetLanguage?: string | null;
  status?: CampaignUpdateStatus;
}

export interface Keyword {
  id: number;
  campaignId: number;
  /** @nullable */
  campaignName?: string | null;
  /** @nullable */
  clientName?: string | null;
  keywordText: string;
  /** @nullable */
  currentRank?: number | null;
  /** @nullable */
  previousRank?: number | null;
  /** @nullable */
  searchVolume?: number | null;
  /** @nullable */
  rankChange?: number | null;
  lastUpdated: string;
}

export interface KeywordInput {
  campaignId: number;
  keywordText: string;
  currentRank?: number;
  previousRank?: number;
  searchVolume?: number;
}

export interface KeywordUpdate {
  keywordText?: string;
  /** @nullable */
  currentRank?: number | null;
  /** @nullable */
  previousRank?: number | null;
  /** @nullable */
  searchVolume?: number | null;
}

export interface KeywordBulkInput {
  campaignId: number;
  keywords: string[];
}

export interface RankHistoryPoint {
  date: string;
  /** @nullable */
  rank: number | null;
}

export type BacklinkStatus = typeof BacklinkStatus[keyof typeof BacklinkStatus];


export const BacklinkStatus = {
  new: 'new',
  lost: 'lost',
} as const;

export interface Backlink {
  id: number;
  clientId: number;
  /** @nullable */
  clientName?: string | null;
  sourceUrl: string;
  targetUrl: string;
  /** @nullable */
  anchorText?: string | null;
  /** @nullable */
  authorityScore?: number | null;
  isToxic?: boolean;
  status: BacklinkStatus;
  firstDetected: string;
  /** @nullable */
  lastSeen?: string | null;
}

export type BacklinkInputStatus = typeof BacklinkInputStatus[keyof typeof BacklinkInputStatus];


export const BacklinkInputStatus = {
  new: 'new',
  lost: 'lost',
} as const;

export interface BacklinkInput {
  clientId: number;
  sourceUrl: string;
  targetUrl: string;
  anchorText?: string;
  authorityScore?: number;
  isToxic?: boolean;
  status?: BacklinkInputStatus;
}

export type BacklinkUpdateStatus = typeof BacklinkUpdateStatus[keyof typeof BacklinkUpdateStatus];


export const BacklinkUpdateStatus = {
  new: 'new',
  lost: 'lost',
} as const;

export interface BacklinkUpdate {
  /** @nullable */
  anchorText?: string | null;
  /** @nullable */
  authorityScore?: number | null;
  isToxic?: boolean;
  status?: BacklinkUpdateStatus;
}

export interface BacklinkBulkInput {
  clientId: number;
  backlinks: BacklinkInput[];
}

export interface BacklinkChartPoint {
  date: string;
  newCount: number;
  lostCount: number;
}

export interface DisavowList {
  content: string;
  count: number;
}

export interface Competitor {
  id: number;
  clientId: number;
  domain: string;
  /** @nullable */
  backlinkCount?: number | null;
  createdAt: string;
}

export interface CompetitorInput {
  clientId: number;
  domain: string;
}

export interface Report {
  id: number;
  clientId: number;
  /** @nullable */
  clientName?: string | null;
  campaignId: number;
  /** @nullable */
  campaignName?: string | null;
  /** @nullable */
  summary?: string | null;
  createdAt: string;
}

export interface ReportInput {
  clientId: number;
  campaignId: number;
}

export type ListClientsParams = {
search?: string;
page?: number;
limit?: number;
};

export type ListCampaignsParams = {
clientId?: number;
status?: string;
};

export type ListKeywordsParams = {
campaignId?: number;
};

export type ListBacklinksParams = {
clientId?: number;
status?: string;
};

export type GetBacklinksChartParams = {
clientId?: number;
};

export type GetDisavowListParams = {
clientId?: number;
};

export type ListCompetitorsParams = {
clientId?: number;
};

export type ListReportsParams = {
clientId?: number;
};