# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SEO Local Admin Panel API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current admin user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          description: Not authenticated

  # Dashboard
...
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List all clients
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientsPage"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create a new client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get a single client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "404":
          description: Not found
    patch:
      operationId: updateClient
      tags: [clients]
      summary: Update a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientUpdate"
      responses:
        "200":
          description: Updated client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Delete a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /clients/{id}/campaigns:
    get:
      operationId: getClientCampaigns
      tags: [clients]
      summary: Get all campaigns for a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client campaigns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Campaign"
  /clients/{id}/keywords:
    get:
      operationId: getClientKeywords
      tags: [clients]
      summary: Get all keywords for a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client keywords
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Keyword"
  /clients/{id}/backlinks:
    get:
      operationId: getClientBacklinks
      tags: [clients]
      summary: Get all backlinks for a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client backlinks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Backlink"

  # Plans
...
  /campaigns:
    get:
      operationId: listCampaigns
      tags: [campaigns]
      summary: List all campaigns
      parameters:
        - name: clientId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of campaigns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Campaign"
    post:
      operationId: createCampaign
      tags: [campaigns]
      summary: Create a campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{id}:
    get:
      operationId: getCampaign
      tags: [campaigns]
      summary: Get a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
    patch:
      operationId: updateCampaign
      tags: [campaigns]
      summary: Update a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignUpdate"
      responses:
        "200":
          description: Updated campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
    delete:
      operationId: deleteCampaign
      tags: [campaigns]
      summary: Delete a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /campaigns/{id}/keywords:
    get:
      operationId: getCampaignKeywords
      tags: [campaigns]
      summary: Get all keywords for a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign keywords
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Keyword"

  # Keywords
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/AdminUser"
    AdminUser:
      type: object
      required: [id, email]
      properties:
        id:
          type: integer
        email:
          type: string
        createdAt:
          type: string
...
    Client:
      type: object
      required: [id, name, email, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        company:
          type: ["string", "null"]
        websiteUrl:
          type: ["string", "null"]
        assignedPlanId:
          type: ["integer", "null"]
        planName:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive]
        createdAt:
          type: string
        updatedAt:
          type: string
    ClientInput:
      type: object
      required: [name, email]
      properties:
        name:
          type: string
        email:
          type: string
        company:
          type: string
        websiteUrl:
          type: string
        assignedPlanId:
          type: integer
        status:
          type: string
          enum: [active, inactive]
    ClientUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        company:
          type: string
        websiteUrl:
          type: string
        assignedPlanId:
          type: ["integer", "null"]
        status:
          type: string
          enum: [active, inactive]
    ClientsPage:
      type: object
      required: [clients, total, page, limit]
      properties:
        clients:
          type: array
          items:
            $ref: "#/components/schemas/Client"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...
    Campaign:
      type: object
      required: [id, clientId, name, targetDomain, status, createdAt]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        clientName:
          type: ["string", "null"]
        name:
          type: string
        targetDomain:
          type: string
        targetLocation:
          type: ["string", "null"]
        targetLanguage:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, paused, completed]
        keywordCount:
          type: integer
        createdAt:
          type: string
    CampaignInput:
      type: object
      required: [clientId, name, targetDomain]
      properties:
        clientId:
          type: integer
        name:
          type: string
        targetDomain:
          type: string
        targetLocation:
          type: string
        targetLanguage:
          type: string
        status:
          type: string
          enum: [active, paused, completed]
    CampaignUpdate:
      type: object
      properties:
        name:
          type: string
        targetDomain:
          type: string
        targetLocation:
          type: ["string", "null"]
        targetLanguage:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, paused, completed]
    Keyword:
      type: object
      required: [id, campaignId, keywordText, lastUpdated]
      properties:
        id:
          type: integer
        campaignId:
          type: integer
        campaignName:
          type: ["string", "null"]
        clientName:
          type: ["string", "null"]
        keywordText:
          type: string
        currentRank:
          type: ["integer", "null"]
        previousRank:
          type: ["integer", "null"]
        searchVolume:
          type: ["integer", "null"]
        rankChange:
          type: ["integer", "null"]
        lastUpdated:
          type: string
...
    Backlink:
      type: object
      required: [id, clientId, sourceUrl, targetUrl, status, firstDetected]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        clientName:
          type: ["string", "null"]
        sourceUrl:
          type: string
        targetUrl:
          type: string
        anchorText:
          type: ["string", "null"]
        authorityScore:
          type: ["integer", "null"]
        isToxic:
          type: boolean
        status:
          type: string
          enum: [new, lost]
        firstDetected:
          type: string
        lastSeen:
          type: ["string", "null"]
...