# lib/api-zod/src/generated/types/clientsPage.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SEO Local Admin Panel API
 * OpenAPI spec version: 0.1.0
 */
import type { Client } from './client';

export interface ClientsPage {
  clients: Client[];
  total: number;
  page: number;
  limit: number;
}